[feat] 단건 조회 구현

+++ assignment_3rd/src/main/java/com/example/assignment_3rd/board/repository/BoardRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface BoardRepository {
     Board create(Board board);
-    Optional<Board> findById(int id);
+    Optional<Board> findById(Long id);
     List<Board> findAll();
     Board update(Board board);
     Long delete(Long id);
